DEV-5664 an int is sufficient here
git-svn-id: http://svn.terracotta.org/svn/ehcache@3902 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/cachemanager_pools/src/main/java/net/sf/ehcache/config/MemoryUnit.java
@@ -40,10 +40,10 @@ public enum MemoryUnit {
         }
     }
 
-    private        final long offset;
+    private        final int offset;
     private        final char unit;
 
-    MemoryUnit(final long offset, final char unit) {
+    MemoryUnit(final int offset, final char unit) {
         this.offset = offset;
         this.unit = unit;
     }
